UPDATE | Stripe Form rendered on the page with header/footer
